+++ gsoap/wsdl/wsdl2h.cpp
@@ -5,7 +5,7 @@
 
 --------------------------------------------------------------------------------
 gSOAP XML Web services tools
-Copyright (C) 2000-2019, Robert van Engelen, Genivia Inc. All Rights Reserved.
+Copyright (C) 2000-2020, Robert van Engelen, Genivia Inc. All Rights Reserved.
 This software is released under one of the following licenses:
 GPL or Genivia's license for commercial use.
 --------------------------------------------------------------------------------
@@ -127,12 +127,12 @@ const char sizeparaformat[]            = "  $ %-35s __size%s%s";
 const char pointertemplateparaformat[] = "    %s<%-21s>*%s%s%s";
 const char derivedformat[]             = "  [ %-35s *%-30s; ]";
 
-const char copyrightnotice[] = "\n**  The gSOAP WSDL/WADL/XSD processor for C and C++, wsdl2h release " WSDL2H_VERSION "\n**  Copyright (C) 2000-2019 Robert van Engelen, Genivia Inc.\n**  All Rights Reserved. This product is provided \"as is\", without any warranty.\n**  The wsdl2h tool and its generated software are released under the GPL.\n**  ----------------------------------------------------------------------------\n**  A commercial use license is available from Genivia Inc., contact@genivia.com\n**  ----------------------------------------------------------------------------\n\n";
+const char copyrightnotice[] = "\n**  The gSOAP WSDL/WADL/XSD processor for C and C++, wsdl2h release " WSDL2H_VERSION "\n**  Copyright (C) 2000-2020 Robert van Engelen, Genivia Inc.\n**  All Rights Reserved. This product is provided \"as is\", without any warranty.\n**  The wsdl2h tool and its generated software are released under the GPL.\n**  ----------------------------------------------------------------------------\n**  A commercial use license is available from Genivia Inc., contact@genivia.com\n**  ----------------------------------------------------------------------------\n\n";
 
 const char licensenotice[]   = "\
 --------------------------------------------------------------------------------\n\
 gSOAP XML Web services tools\n\
-Copyright (C) 2000-2019, Robert van Engelen, Genivia Inc. All Rights Reserved.\n\
+Copyright (C) 2000-2020, Robert van Engelen, Genivia Inc. All Rights Reserved.\n\
 The wsdl2h tool and its generated software are released under the GPL.\n\
 This software is released under the GPL with the additional exemption that\n\
 compiling, linking, and/or using OpenSSL is allowed.\n\
@@ -584,6 +584,7 @@ static void options(int argc, char **argv)
 -z7     compatibility up to 2.8.59: don't generate std::vector of class of union\n\
 -z8     compatibility up to 2.8.74: don't gen quals for doc/lit wrapped patterns\n\
 -z9     compatibility up to 2.8.93: always qualify element/attribute refs\n\
+-z10    compatibility up to 2.8.96: gen quals even when defined w/o namespace\n\
 -_      don't generate _USCORE (replace with Unicode code point _x005f)\n\
 infile.wsdl infile.xsd http://www... list of input sources (if none reads stdin)\n\
 \n");